[Refactor/#135] FaqScreen을 screens.faq 패키지로 이동
