+++ feature/my/src/main/java/com/ilsangtech/ilsang/feature/my/navigation/MyTabNavigation.kt
@@ -7,13 +7,13 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
-import com.ilsangtech.ilsang.feature.my.FaqScreen
 import com.ilsangtech.ilsang.feature.my.MyChallengeScreen
 import com.ilsangtech.ilsang.feature.my.MyTitleScreen
 import com.ilsangtech.ilsang.feature.my.TermsScreen
 import com.ilsangtech.ilsang.feature.my.UserProfileEditScreen
 import com.ilsangtech.ilsang.feature.my.WithdrawalScreen
 import com.ilsangtech.ilsang.feature.my.component.CustomerCenterScreen
+import com.ilsangtech.ilsang.feature.my.screens.faq.FaqScreen
 import com.ilsangtech.ilsang.feature.my.screens.mytab.MyTabScreen
 import com.ilsangtech.ilsang.feature.my.screens.setting.SettingScreen
 import kotlinx.serialization.Serializable
